scm_optimize
